requirements.txt
chroma-hnswlib==0.7.2
fastapi>=0.95.2, <0.100.0
graphlib_backport==1.0.3; python_version < '3.9'
importlib-resources
numpy==1.21.6
onnxruntime==1.14.1
overrides==7.3.1
posthog==2.4.0
pulsar-client==3.1.0
pydantic>=1.9,<2.0
pypika==0.48.9
requests==2.28.1
tokenizers==0.13.2
tqdm==4.65.0
typing_extensions==4.5.0
uvicorn[standard]==0.18.3

# Additional dependencies
openai
llama-index
streamlit
python-dotenv
langchain
pypdf
chromadb
docx2txt
sentence-transformers
PyYAML